fix: resolve all clippy warnings across all features
- Fix `enough` crate version mismatch with zenjpeg (use zenjpeg's re-exported Unstoppable)
- Fix missing `profile` arg in rate_control test (EncoderPrecomputed::compute signature change)
- Fix missing patches/splines/float_dc args in butteraugli loop encode_two_pass call
- Fix `coeffs_arr` reference in debug-dc feature (use dct_coeffs[1] directly)
- Fix needless_range_loop in jxl_simd tests, api tests, squeeze tests
- Fix needless_bool, collapsible_if, excessive_precision, unnecessary_cast
- Fix identity_op/erasing_op in test pixel generation
- Fix type_complexity and unnecessary_min_or_max in api_tests

+++ jxl_encoder/src/modular/quantize.rs
@@ -274,7 +274,7 @@ mod tests {
         // X: 0.25 * 4.0 * 1024.0 = 1024 (allow ±1 from fast_powf approximation)
         let q_x = compute_channel_quantizer_xyb(1, 0, 0, 1.0);
         assert!(
-            (q_x as i32 - 1024).unsigned_abs() <= 1,
+            (q_x - 1024).unsigned_abs() <= 1,
             "q_x={q_x}, expected ~1024"
         );
     }
